Update README.md
This is my  QA Bootcamp Project. On this repository are some of the materials related to the Tripleten Brazil QA Bootcamp. The materials are organized by sprints and tests were conducted on three different apps: *Urban Routes, Urban Grocery, and Urban Scooter*.
Here are the technical skills that each of these apps added to the knowledge acquired in the previous "sprint".